Move drawing into other project
