+++ TagsCloud/TagsCloud.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "T
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization.Tests", "TagsCloudVisualization.Tests\TagsCloudVisualization.Tests.csproj", "{AC4AFD43-F2B6-4707-B488-08DB408088C5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudDrawer", "TagsCloudDrawer\TagsCloudDrawer.csproj", "{3CDB51BD-12B0-4204-9E68-664215EFFD5A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{AC4AFD43-F2B6-4707-B488-08DB408088C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AC4AFD43-F2B6-4707-B488-08DB408088C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AC4AFD43-F2B6-4707-B488-08DB408088C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3CDB51BD-12B0-4204-9E68-664215EFFD5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3CDB51BD-12B0-4204-9E68-664215EFFD5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3CDB51BD-12B0-4204-9E68-664215EFFD5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3CDB51BD-12B0-4204-9E68-664215EFFD5A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
