+++ TagsCloud/TagsCloudDrawer/ImageSettings/IImageSettingsProvider.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+
+namespace TagsCloudDrawer.ImageSettings
+{
+    public interface IImageSettingsProvider
+    {
+        Color BackgroundColor { get; }
+        Size ImageSize { get; }
+    }
+}
