+++ TagsCloud/TagsCloudDrawer/TagsCloudDrawer.csproj
@@ -1,17 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <OutputType>Exe</OutputType>
         <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
         <PackageReference Include="System.Drawing.Common" Version="6.0.0-rc.2.21480.5"/>
-        <PackageReference Include="System.Drawing.Primitives" Version="4.3.0"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\TagsCloudDrawer\TagsCloudDrawer.csproj"/>
     </ItemGroup>
 
 </Project>
