+++ TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/DrawerSettingsProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using TagsCloudVisualization.ColorGenerators;
+using TagsCloudDrawer.ColorGenerators;
 
-namespace TagsCloudVisualization.TagsCloudDrawer.TagsCloudDrawerSettingsProvider
+namespace TagsCloudVisualization.DrawerSettingsProvider
 {
     public class DrawerSettingsProvider : IDrawerSettingsProvider
     {
