--- TagsCloud/TagsCloudVisualization/ImageCreator/IImageSettingsProvider.cs
@@ -1,10 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization.ImageCreator
-{
-    public interface IImageSettingsProvider
-    {
-        public Color BackgroundColor { get; init; }
-        public Size ImageSize { get; init; }
-    }
-}
